Keep the spec importer under the size and complexity ratchets
The multi-segment path parameter pushed `spec_import.py` past 600 lines and
`_build_operation` past a complexity of 15. Rather than raise the baseline --
which records pre-existing debt, not permission for new code -- the pure
spec-walking helpers move to `spec_helpers.py`, which is the split the file's
own section comment already described: resolving refs and picking content
schemas is a separate concern from building descriptors out of them. Building
the execution descriptor becomes its own function for the same reason.

No caller imported those helpers by name, and regenerating the GitHub catalog
produces a byte-identical file.
