+++ lemma-backend/app/modules/connectors/infrastructure/openapi/spec_import.py
@@ -15,180 +15,23 @@ by ``OpenApiHttpExecutor``.
 
 from __future__ import annotations
 
-import copy
-import re
 from dataclasses import dataclass, field
 from typing import Any
 
+from app.modules.connectors.infrastructure.openapi.spec_helpers import (
+    _BODY_PREFERRED_TYPES,
+    _HTTP_METHODS,
+    _RESPONSE_PREFERRED_TYPES,
+    build_parameter_entry,
+    build_tool_name,
+    pick_content_schema,
+    pick_success_response,
+    prefers_binary_response,
+    resolve_once,
+)
+
 # --- constants --------------------------------------------------------------
 
-SUCCESS_RESPONSE_CODES = ("200", "201", "202", "203", "204", "206")
-
-# Provider-generic noise params (Google-style) that are auth/formatting knobs,
-# never real operation inputs. Harmless to drop for any provider.
-IGNORED_PARAMETER_NAMES = {
-    "access_token",
-    "alt",
-    "callback",
-    "key",
-    "oauth_token",
-    "prettyPrint",
-    "quotaUser",
-    "uploadType",
-    "upload_protocol",
-    "userIp",
-    "$.xgafv",
-}
-
-_HTTP_METHODS = {"GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS"}
-
-# Multipart-/binary-aware order — we must NOT collapse everything to JSON.
-_BODY_PREFERRED_TYPES = [
-    "application/json",
-    "multipart/form-data",
-    "application/octet-stream",
-    "*/*",
-]
-_RESPONSE_PREFERRED_TYPES = ["application/json", "*/*"]
-
-
-# --- ported pure helpers ----------------------------------------------------
-
-
-def resolve_ref(spec: dict[str, Any], ref: str) -> Any:
-    node: Any = spec
-    for part in ref.removeprefix("#/").split("/"):
-        node = node[part]
-    return node
-
-
-def resolve_once(spec: dict[str, Any], value: Any) -> Any:
-    if isinstance(value, dict) and "$ref" in value:
-        return resolve_ref(spec, value["$ref"])
-    return value
-
-
-def deep_resolve_refs(
-    spec: dict[str, Any],
-    value: Any,
-    *,
-    seen_refs: set[str] | None = None,
-) -> Any:
-    seen_refs = seen_refs or set()
-    value = copy.deepcopy(value)
-    if isinstance(value, dict):
-        if "$ref" in value:
-            ref = value["$ref"]
-            if ref in seen_refs:
-                return {"$ref": ref}
-            return deep_resolve_refs(
-                spec,
-                resolve_ref(spec, ref),
-                seen_refs={*seen_refs, ref},
-            )
-        return {
-            key: deep_resolve_refs(spec, item, seen_refs=seen_refs)
-            for key, item in value.items()
-        }
-    if isinstance(value, list):
-        return [deep_resolve_refs(spec, item, seen_refs=seen_refs) for item in value]
-    return value
-
-
-def normalize_schema(schema: dict[str, Any] | None) -> dict[str, Any]:
-    if not schema:
-        return {"type": "object", "additionalProperties": True}
-    return schema
-
-
-def build_parameter_entry(
-    spec: dict[str, Any], parameter: dict[str, Any]
-) -> dict[str, Any] | None:
-    parameter = resolve_once(spec, parameter)
-    name = parameter["name"]
-    if name in IGNORED_PARAMETER_NAMES:
-        return None
-    raw_schema = parameter.get("schema") or {}
-    return {
-        "name": name,
-        "location": parameter["in"],
-        "required": parameter.get("required", False),
-        "description": parameter.get("description"),
-        "schema": normalize_schema(deep_resolve_refs(spec, raw_schema)),
-        "style": parameter.get("style"),
-        "explode": parameter.get("explode"),
-    }
-
-
-def pick_content_schema(
-    spec: dict[str, Any],
-    content: dict[str, Any] | None,
-    *,
-    preferred_types: list[str] | None = None,
-) -> tuple[str, dict[str, Any], str | None]:
-    if not content:
-        return "application/json", {"type": "object", "additionalProperties": True}, None
-    preferred = preferred_types or ["application/json", "*/*"]
-    for content_type in preferred:
-        if content_type in content:
-            item = resolve_once(spec, content[content_type])
-            raw_schema = item.get("schema") or {}
-            schema_ref = raw_schema.get("$ref") if isinstance(raw_schema, dict) else None
-            return content_type, normalize_schema(deep_resolve_refs(spec, raw_schema)), schema_ref
-    first_type, first_value = next(iter(content.items()))
-    item = resolve_once(spec, first_value)
-    raw_schema = item.get("schema") or {}
-    schema_ref = raw_schema.get("$ref") if isinstance(raw_schema, dict) else None
-    return first_type, normalize_schema(deep_resolve_refs(spec, raw_schema)), schema_ref
-
-
-def build_tool_name(operation_id: str, method: str, path: str) -> str:
-    if operation_id:
-        parts = []
-        for chunk in operation_id.replace("/", ".").split("."):
-            chunk = re.sub(r"([a-z0-9])([A-Z])", r"\1_\2", chunk)
-            chunk = re.sub(r"[^a-zA-Z0-9]+", "_", chunk).strip("_").lower()
-            if chunk:
-                parts.append(chunk)
-        return "_".join(parts)
-    slug = path.strip("/").replace("/", "_").replace("{", "").replace("}", "")
-    return f"{method.lower()}_{slug}"
-
-
-def pick_success_response(operation: dict[str, Any]) -> dict[str, Any] | None:
-    responses = operation.get("responses", {}) or {}
-    for status_code in SUCCESS_RESPONSE_CODES:
-        response = responses.get(status_code)
-        if response is not None:
-            return response
-    for status_code, response in responses.items():
-        if str(status_code).startswith("2"):
-            return response
-    return None
-
-
-def prefers_binary_response(
-    *,
-    operation_id: str,
-    path: str,
-    content: dict[str, Any] | None,
-) -> bool:
-    marker = f"{operation_id} {path}".lower()
-    keyword_match = any(
-        token in marker
-        for token in ("download", "export", "thumbnail", "avatar", "image", "tarball", "zipball")
-    )
-    if not content:
-        return keyword_match
-    media_types = {str(item).split(";", 1)[0].strip().lower() for item in content}
-    has_binary_media = any(
-        media_type in {"*/*", "application/octet-stream"}
-        or media_type.startswith("image/")
-        or media_type.startswith("audio/")
-        or media_type.startswith("video/")
-        for media_type in media_types
-    )
-    return keyword_match and has_binary_media
 
 
 # --- descriptor building ----------------------------------------------------
@@ -439,6 +282,44 @@ def _resolve_response(
     return False, output_schema
 
 
+def _build_execution(
+    *,
+    method: str,
+    path: str,
+    server_url: str,
+    params: "_Parameters",
+    request_body_desc: dict[str, Any] | None,
+    binary: bool,
+    override: dict[str, Any],
+    default_headers: dict[str, str] | None,
+) -> dict[str, Any]:
+    """The descriptor ``OpenApiHttpExecutor`` replays to make the call."""
+    execution: dict[str, Any] = {
+        "kind": "http",
+        "mode": "openapi",
+        "method": method,
+        "path": path,
+        "server_url": server_url,
+        "path_params": params.path_params,
+        "query_params": params.query_params,
+        "header_params": params.header_params,
+        "request_body": request_body_desc,
+        "response": {"binary": binary},
+    }
+    # Path parameters whose value spans several URL segments (a git ref such as
+    # `heads/main`). A spec cannot express this, so it comes from the override.
+    multi_segment = [
+        name
+        for name in override.get("multi_segment_path_params") or []
+        if name in params.path_params
+    ]
+    if multi_segment:
+        execution["multi_segment_path_params"] = multi_segment
+    if default_headers:
+        execution["default_headers"] = dict(default_headers)
+    return execution
+
+
 def _build_operation(
     spec: dict[str, Any],
     *,
@@ -482,29 +363,16 @@ def _build_operation(
     if required:
         input_schema["required"] = required
 
-    execution: dict[str, Any] = {
-        "kind": "http",
-        "mode": "openapi",
-        "method": method,
-        "path": path,
-        "server_url": server_url,
-        "path_params": params.path_params,
-        "query_params": params.query_params,
-        "header_params": params.header_params,
-        "request_body": request_body_desc,
-        "response": {"binary": binary},
-    }
-    # Path parameters whose value spans several URL segments (a git ref such as
-    # `heads/main`). A spec cannot express this, so it comes from the override.
-    multi_segment = [
-        name
-        for name in override.get("multi_segment_path_params") or []
-        if name in params.path_params
-    ]
-    if multi_segment:
-        execution["multi_segment_path_params"] = multi_segment
-    if default_headers:
-        execution["default_headers"] = dict(default_headers)
+    execution = _build_execution(
+        method=method,
+        path=path,
+        server_url=server_url,
+        params=params,
+        request_body_desc=request_body_desc,
+        binary=binary,
+        override=override,
+        default_headers=default_headers,
+    )
 
     return OpenAPIOperation(
         public_name=public_name,
